refactor(backend): split control plane desired version controller
Split the combined control plane desired-version controller into two
controllers, each gated by a NeedsWork method on the cluster's active
versions:

- controlPlaneInitialVersionSyncer ("ControlPlaneInitialVersion"): seeds
  the initial desired version before any active versions exist (Case 1).
  NeedsWork is true when ActiveVersions is empty.
- controlPlaneUpgradeVersionSyncer ("ControlPlaneDesiredVersion"):
  handles z-stream and y-stream upgrades plus the downgrade guard
  (Cases 2+3). NeedsWork is true when ActiveVersions is non-empty.

Shared state and behaviour (the create-race gate
shouldDetermineDesiredVersion, Cincinnati client construction, the
desired-version persistence/IntentFailed policy, and the Cincinnati
version-selection free functions) live in desiredVersionSyncerCommon,
embedded by both syncers.

The upgrade controller keeps the "ControlPlaneDesiredVersion" name and
the existing NewControlPlaneDesiredVersionController constructor for
metrics continuity and to avoid touching backend.go wiring. The new
NewControlPlaneInitialVersionController is intentionally not wired yet;
that is left to a follow-up.

Tests are split accordingly: initial-version tests in the initial
controller test file, upgrade tests in the upgrade controller test file,
and shared fixtures plus the gate test in a common test file.
